% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iter.R
\name{iter}
\alias{iter}
\title{Create an Iterable Object}
\usage{
iter(x)
}
\arguments{
\item{x}{object to be looped across}
}
\value{
a vector
}
\description{
Vector \code{\link{comprehension}} iterates over an object, but the default
behavior may not be desirable for custom classes. \code{iter} allows the
user to specify how the object behaves within a comprehension, or other loop
in the \code{eList} package. Unless a method is specified for an object,
iter will attempt to convert it to a list except for atomic vectors.
}
\examples{
e <- new.env()
e$x <- 10
e$y <- letters[1:10]
iter(e)

}
