% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tools.R
\name{null.omit}
\alias{null.omit}
\title{Remove 'NULL' Entries from List}
\usage{
null.omit(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
\code{x} without \code{NULL} entries
}
\description{
Function removes all items that are \code{NULL} or empty from a list
or other object.
}
\examples{
l <- list(a=2, b=NULL, c = 3)
length(l) == 3

k <- null.omit(l)
length(k) == 2

}
